Update compress.cpp
I plugged pretty much the entire program into a function called compressData(). I feel that this wasn't exactly what 'bminor' wanted me to do though. compressData() takes in a stream of data to be compressed, and returns another stream of compressed data after it was already written to an output file.